removed stack from XDK
