+++ drivers/x540-nic/Makefile
@@ -1,36 +1,25 @@
 include ../../mk/global.mk
 BOOST_LIBS = -lboost_thread
-CXXFLAGS += -g -O3 -std=c++11 -Wall -DCONFIG_BUILD_DEBUG -I$(XDK_BASE)/lib/libexo -I$(XDK_BASE)/lib/libexo/exo/ -I$(XDK_BASE)/drivers/x540-nic/ -I$(XDK_BASE)/net -I$(XDK_BASE)/net/micro-udp -I$(XDK_BASE) -I$(XDK_BASE)/include -I$(XDK_BASE)/lib/libcommon -I$(XDK_BASE)/lib/libcomponent/component -I$(XDK_BASE)/lib/libcomponent/
+CXXFLAGS += -g -O3 -std=c++11 -Wall -DCONFIG_BUILD_DEBUG -I$(XDK_BASE)/lib/libexo -I$(XDK_BASE)/lib/libexo/exo/ -I$(XDK_BASE)/drivers/x540-nic/ -I$(XDK_BASE) -I$(XDK_BASE)/include -I$(XDK_BASE)/lib/libcommon -I$(XDK_BASE)/lib/libcomponent/component -I$(XDK_BASE)/lib/libcomponent/
 LIBS = $(XDK_LIBS) $(BOOST_LIBS) -L$(XDK_BASE)/lib/libexo -L$(XDK_BASE)/lib/libcomponent -lexo -lcomponent -lpthread -lnuma -ldl -lrt -lm -ltinyxml  
 
 NIC_COMPONENT_SRC := nic_component.cc \
                      ./x540/x540_device.cc \
-                     ./x540/xml_config_parser.cc \
-                     $(XDK_BASE)/net/micro-udp/stack/exo_stack.cc
+                     xml_config_parser.cc \
 
 MEM_COMPONENT_SRC := mem_component.cc \
-                     ./x540/x540_device.cc \
-                     ./x540/xml_config_parser.cc \
-                    $(XDK_BASE)/net/micro-udp/stack/exo_stack.cc
+                     xml_config_parser.cc \
 
-STACK_COMPONENT_SRC := $(XDK_BASE)/net/micro-udp/stack_component.cc \
-                       ./x540/x540_device.cc \
-                       ./x540/xml_config_parser.cc \
-                       $(XDK_BASE)/net/micro-udp/stack/exo_stack.cc
-
-all: nic_component stack_component mem_component app_main
+all: nic_component mem_component app_main
 
 app_main:
-	g++ $(CXXFLAGS) -o $@ main.cc ./x540/xml_config_parser.cc $(LIBS) -Wl,-rpath=$(XDK_BASE)/lib/libexo 
+	g++ $(CXXFLAGS) -o $@ main.cc xml_config_parser.cc $(LIBS) -Wl,-rpath=$(XDK_BASE)/lib/libexo 
 
 nic_component: 
 	g++ $(CXXFLAGS) -g -shared -fPIC -Wl,-soname,libcomp_nic.so.1 -o libcomp_nic.so.1 $(NIC_COMPONENT_SRC) $(LIBS)
 
 mem_component: 
 	g++ $(CXXFLAGS) -g -shared -fPIC -Wl,-soname,libcomp_mem.so.1 -o libcomp_mem.so.1 $(MEM_COMPONENT_SRC) $(LIBS)
-
-stack_component: 
-	g++ $(CXXFLAGS) -g -shared -fPIC -Wl,-soname,libcomp_stack.so.1 -o libcomp_stack.so.1 $(STACK_COMPONENT_SRC) $(LIBS)
 
 clean:
 	rm -f app_main libcomp*.so.1 *.o x540/*.o
