+++ mk/global.mk.in
@@ -1,4 +1,5 @@
-XDK_BASE = $(HOME)/xdk
+XDK_HOME = `pwd`
+XDK_BASE = $(XDK_HOME)
 
 XDK_INCLUDES = -I$(XDK_BASE)/lib/libexo -I$(XDK_BASE)/lib/libcomponent -I$(XDK_BASE)/lib/libcommon
 XDK_LIBS = -L$(XDK_BASE)/lib/libexo -L$(XDK_BASE)/lib/libcomponent -L$(XDK_BASE)/lib/libcommon -lcomponent -lexo -lcommon
