--- net/README.txt
@@ -1,1 +1,0 @@
-This directory contains implementations of network protocol stacks.
